microservices/reports/pyproject.toml
[tool.poetry]
name = "reports"
version = "0.1.0"
description = ""
authors = ["hfmartinez <heberthfabianmartinez@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
opencv-python = "^4.9.0.80"
sqlalchemy = "^2.0.28"
fastapi = "^0.110.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
